+++ src/components/common/shop-info/shop-info-card.tsx
